+++ spec/nim/tcast_to_top.nim
@@ -6,6 +6,6 @@ import auxiliary/test_utils
 
 let r = CastToTop.fromFile("src/fixed_struct.bin")
 
-assert r.code == uint8(80)
-assert r.header.code == uint8(65)
-assert r.headerCasted.code == uint8(65)
+assert r.code == 80
+assert r.header.code == 65
+assert r.headerCasted.code == 65
