+++ spec/nim/tdefault_bit_endian_mod.nim
@@ -6,7 +6,7 @@ import auxiliary/test_utils
 
 let r = DefaultBitEndianMod.fromFile("src/fixed_struct.bin")
 
-assert r.main.one == uint64(336)
-assert r.main.two == uint64(8608)
-assert r.main.nest.two == uint64(11595)
-assert r.main.nestBe.two == uint64(12799)
+assert r.main.one == 336
+assert r.main.two == 8608
+assert r.main.nest.two == 11595
+assert r.main.nestBe.two == 12799
