+++ spec/nim/tnav_parent_override.nim
@@ -6,6 +6,6 @@ import auxiliary/test_utils
 
 let r = NavParentOverride.fromFile("src/nav_parent_codes.bin")
 
-assert r.childSize == uint8(3)
+assert r.childSize == 3
 assert r.child1.data == @[73'u8, 49'u8, 50'u8]
 assert r.mediator2.child2.data == @[51'u8, 66'u8, 98'u8]
