+++ spec/nim/tnav_parent_switch.nim
@@ -6,6 +6,6 @@ import auxiliary/test_utils
 
 let r = NavParentSwitch.fromFile("src/nav_parent_switch.bin")
 
-assert r.category == uint8(1)
-assert (NavParentSwitch_Element1(r.content)).foo == uint8(66)
-assert (NavParentSwitch_Element1(r.content)).subelement.bar == uint8(255)
+assert r.category == 1
+assert (NavParentSwitch_Element1(r.content)).foo == 66
+assert (NavParentSwitch_Element1(r.content)).subelement.bar == 255
